GithubLinkFinder/GithubLinkFinder/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.12",
    "discord-py>=2.4.0",
    "flask>=3.1.0",
    "psutil>=6.1.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "twilio>=9.4.5",
]
